+++ client/src/pages/CollectionPage/ui/CollectionHeader/CollectionHeader.tsx
@@ -8,31 +8,24 @@ import { observer } from 'mobx-react-lite';
 import { Collection } from 'entities/Collection';
 import Markdown from 'react-markdown';
 import { itemState } from 'entities/Item';
+import { PageHeader } from 'shared/ui/PageHeader/PageHeader';
+import { ElementsTypes } from 'shared/class/ElementState';
 import cls from './CollectionHeader.module.scss';
 
 const { Title } = Typography;
 
 interface CollectionHeaderProps {
   className?: string
-  collection: Partial<Collection>
+  collection: Collection
 }
 
 export const CollectionHeader: FC<CollectionHeaderProps> = observer((props) => {
     const { collection } = props;
     const { t } = useTranslation();
 
-    const handleClick = () => {
-        itemState.setValues(null);
-        itemState.openModal();
-    };
-
     return (
-        <div className={cls.container}>
-            <Avatar
-                shape="square"
-                className={cls.avatar}
-            />
-            <div className={cls.content}>
+        <PageHeader img="" type={ElementsTypes.item} userId={collection.userId}>
+            <>
                 <Title>{collection.title}</Title>
                 <Title level={3}>
                     Theme:
@@ -40,19 +33,7 @@ export const CollectionHeader: FC<CollectionHeaderProps> = observer((props) => {
                     {(collection.theme as any).title}
                 </Title>
                 <Markdown>{collection.description}</Markdown>
-
-            </div>
-            <div className={cls.content}>
-                {(collection.userId === userState.userId || userState.isAdmin) && (
-                    <Button
-                        className={cls.buttons}
-                        onClick={handleClick}
-                    >
-                        +
-                        {t('item')}
-                    </Button>
-                )}
-            </div>
-        </div>
+            </>
+        </PageHeader>
     );
 });
