+++ client/src/pages/UserPage/ui/UserHeader/UserHeader.tsx
@@ -7,6 +7,8 @@ import { useTranslation } from 'react-i18next';
 import { User, userState } from 'entities/User';
 import { observer } from 'mobx-react-lite';
 import { collectionState } from 'entities/Collection';
+import { PageHeader } from 'shared/ui/PageHeader/PageHeader';
+import { ElementsTypes } from 'shared/class/ElementState';
 import cls from './UserHeader.module.scss';
 
 const { Title } = Typography;
@@ -20,30 +22,13 @@ export const UserHeader: FC<UserHeaderProps> = observer((props) => {
     const { user } = props;
     const { t } = useTranslation();
 
-    const handleClick = () => {
-        collectionState.setValues(null);
-        collectionState.openModal();
-    };
-
     return (
-        <div className={cls.container}>
-            <Avatar
-                shape="square"
-                className={cls.avatar}
-            />
-            <div className={cls.content}>
+        <PageHeader img="" type={ElementsTypes.collection} userId={user.id}>
+            <>
                 <Title>{user.username}</Title>
                 <AppStatistic itemsAmount={200} collectionsAmount={300} />
-                {(user.id === userState.userId || userState.isAdmin) && (
-                    <Button
-                        className={cls.buttons}
-                        onClick={handleClick}
-                    >
-                        +
-                        {t('collection')}
-                    </Button>
-                )}
-            </div>
-        </div>
+            </>
+        </PageHeader>
+
     );
 });
