+++ client/src/shared/ui/PageHeader/PageHeader.module.scss
@@ -1,0 +1,24 @@
+.container {
+    display: flex;
+    gap: 50px;
+    align-items: center;
+    position: relative;
+
+    @media (max-width: 700px) {
+        gap: 20px;
+        flex-direction: column;
+        align-items: center;
+    }
+}
+
+.avatar,
+.icon {
+    min-width: 250px;
+    min-height: 250px;
+}
+
+.buttons {
+    position: absolute;
+    right: 0;
+    bottom: 0;
+}
